requirements.txt
flask
kaggle
matplotlib
numpy
opencv-python-headless
pillow
scikit-learn
streamlit
tensorflow
